refactoring AttributeGeneratorDialogsResult (Cause : Too much nested Conditional) and small fix
