+++ src/tool/clients/fmmlxdiagrams/dialogs/results/instancegenerator/AttributeGeneratorNormalDistributionDialogResult.java
@@ -1,0 +1,7 @@
+package tool.clients.fmmlxdiagrams.dialogs.results.instancegenerator;
+
+import tool.clients.fmmlxdiagrams.dialogs.results.DialogResult;
+
+public class AttributeGeneratorNormalDistributionDialogResult extends DialogResult {
+
+}
